add glm4v openai server
